+++ src/wxtiles.ts
@@ -7,7 +7,7 @@ import L from 'leaflet';
 import { WXLOG } from './wxtools';
 import { WxTilesLayer, WxTilesLayerSettings } from './tilesLayer';
 
-export { WxTilesLogging, WxTilesLibSetup, WxGetColorStyles } from './wxtools';
+export { WxTilesLogging, WxTilesLibSetup, WxGetColorStyles, WxGetColorSchemes } from './wxtools';
 export type { ColorStylesWeakMixed, ColorStyleStrict, ColorStylesStrict, Units, ColorSchemes, LibSetupObject } from './wxtools';
 export type { WxTilesLayerSettings, DataSource, Meta, WxTileInfo } from './tilesLayer';
 export { WxTilesLayer } from './tilesLayer';