+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Write your changes below this section. `npm version [major|minor|patch]` will au
 
 ## [Unreleased]
 
+### ADD
+
+- export WxGetColorSchemes
+
 ## [1.0.12] - 2021-12-01
 
 ### REVERT